momentum spread import and test updated
