+++ injector_model/injector_model_full_sc_integral.py
@@ -599,88 +599,3 @@ class InjectorChain_v2:
             df_SC_and_max_intensity.to_csv("output_csv/{}_for_paper.csv".format(output_name), index=True)
             
         return df_all_ions
-    
-    
-    
-class Momentum_Spread(InjectorChain_v2):
-    """"Beam parameter class to calculate typical momentum spread values"""
-
-    def __init__(self):
-        InjectorChain_v2.__init__(self)
-
-    def calculate_sigma_delta_LEIR_for_all_ions(self, num_part=5000)->None:
-        """Find momentum spread for all ions in LEIR"""
-        leir_line = self.line_LEIR_Pb0.copy()
-        leir_line.build_tracker()
-        
-        print('\nLEIR: Calculating momentum spread...')
-        # Iterate over all ion types
-        for i, ion_type in enumerate(self.full_ion_data.columns):
-            self.init_ion(ion_type)
-            
-            LEIR_particle = xp.Particles(mass0 = 1e9 * self.mass_GeV, q0 = self.Q_LEIR, gamma0 = self.LEIR_gamma_inj)
-            leir_line.particle_ref = LEIR_particle
-            
-            particles_leir = xp.generate_matched_gaussian_bunch(
-                num_particles=num_part, 
-                total_intensity_particles=BeamParams_LEIR.Nb,
-                nemitt_x=BeamParams_LEIR.exn, 
-                nemitt_y=BeamParams_LEIR.eyn, 
-                sigma_z= BeamParams_LEIR.sigma_z,
-                particle_ref=leir_line.particle_ref, 
-                line=leir_line)
-            sigma_delta = np.std(particles_leir.delta[particles_leir.state > 0])
-            
-            print('LEIR: for {}, sigma_delta = {}'.format(ion_type, sigma_delta))
-            
-            
-    def calculate_sigma_delta_PS_for_all_ions(self, num_part=5000)->None:
-        """Find momentum spread for all ions in PS"""
-        ps_line = self.line_PS_Pb0.copy()
-        ps_line.build_tracker()
-
-        print('\nPS: Calculating momentum spread...')
-        # Iterate over all ion types
-        for i, ion_type in enumerate(self.full_ion_data.columns):
-            self.init_ion(ion_type)
-            
-            PS_particle = xp.Particles(mass0 = 1e9 * self.mass_GeV, q0 = self.Q_PS, gamma0 = self.PS_gamma_inj)   
-            ps_line.particle_ref = PS_particle
-            
-            particles_ps = xp.generate_matched_gaussian_bunch(
-                num_particles=num_part, 
-                total_intensity_particles=BeamParams_PS.Nb,
-                nemitt_x=BeamParams_PS.exn, 
-                nemitt_y=BeamParams_PS.eyn, 
-                sigma_z= BeamParams_PS.sigma_z,
-                particle_ref=ps_line.particle_ref, 
-                line=ps_line)
-            sigma_delta = np.std(particles_ps.delta[particles_ps.state > 0])
-            
-            print('PS: for {}, sigma_delta = {}'.format(ion_type, sigma_delta))
-            
-
-    def calculate_sigma_delta_SPS_for_all_ions(self, num_part=5000)->None:
-        """Find momentum spread for all ions in SPS"""
-        sps_line = self.line_SPS_Pb0.copy()
-        sps_line.build_tracker()
-
-        print('\nSPS: Calculating momentum spread...')
-        # Iterate over all ion types
-        for i, ion_type in enumerate(self.full_ion_data.columns):
-            self.init_ion(ion_type)
-            
-            SPS_particle = xp.Particles(mass0 = 1e9 * self.mass_GeV, q0 = self.Q_SPS, gamma0 = self.SPS_gamma_inj)
-            sps_line.particle_ref = SPS_particle
-            
-            particles_sps = xp.generate_matched_gaussian_bunch(
-                num_particles=num_part, 
-                total_intensity_particles=BeamParams_SPS.Nb,
-                nemitt_x=BeamParams_SPS.exn, 
-                nemitt_y=BeamParams_SPS.eyn, 
-                sigma_z= BeamParams_SPS.sigma_z,
-                particle_ref=sps_line.particle_ref, 
-                line=sps_line)
-            sigma_delta = np.std(particles_sps.delta[particles_sps.state > 0])
-            
-            print('SPS: for {}, sigma_delta = {}'.format(ion_type, sigma_delta))
